+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## Latest Changes
+
+### Fixed
+
+- Add support for empty batch dimension in `cuequivariance-torch`.
+
 ## 0.1.0 (2024-11-18)
 
 - Beta version of cuEquivariance released.